Added data validation to controllers by adding data annotations to DTOs and route constraints to simple types

+++ api/Controllers/CommentController.cs
@@ -7,6 +7,7 @@ using api.Interfaces;
 using api.Mappers;
 using api.Models;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore.Metadata.Internal;
 
 namespace api.Controllers
 {
@@ -25,15 +26,24 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var comments = await _commentRepo.GetAllAsync();
             var commentsDto = comments.Select(item => item.ToCommentDTO());
 
             return Ok(commentsDto);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")] // the :int is a route constraint
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             Comment? comment = await _commentRepo.GetByIdAsync(id);
             if (comment == null)
             {
@@ -44,9 +54,14 @@ namespace api.Controllers
 
         }
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromBody] CreateCommentRequestDTO commentDto, [FromRoute] int stockId)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             if (!await _stockRepo.StockExists(stockId))
             {
                 return BadRequest("Stock does not exist.");
@@ -58,9 +73,14 @@ namespace api.Controllers
             return CreatedAtAction(nameof(GetById), new { id = commentModel.Id }, commentModel.ToCommentDTO());
 
         }
-        [HttpPut("{commentId}")]
+        [HttpPut("{commentId:int}")]
         public async Task<IActionResult> Update([FromRoute] int commentId, UpdateCommentRequestDTO updateDto)
         {
+            if (!ModelState.IsValid) // model state comes from the controller base class. (also controller base basically says that this is a controller!)
+            {
+                return BadRequest(ModelState);
+            }
+
             var commentModel = await _commentRepo.UpdateComment(commentId, updateDto);
             if (commentModel == null)
             {
@@ -70,9 +90,14 @@ namespace api.Controllers
             return Ok(commentModel.ToCommentDTO());
 
         }
-        [HttpDelete("{commentId}")]
+        [HttpDelete("{commentId:int}")]
         public async Task<IActionResult> Delete([FromRoute] int commentId)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            
             var commentModel = await _commentRepo.DeleteComment(commentId);
             if (commentModel == null)
             {
